drivers: hv: dxgkrnl: Creation of dxgadapter object
Handle creation and destruction of dxgadapter object, which
represents a virtual compute device, projected to the VM by
the host. The dxgadapter object is created when the
corresponding VM bus channel is offered by Hyper-V.

There could be multiple virtual compute device objects,
projected by the host to VM. They are enumerated by
issuing IOCTLs to the /dev/dxg device.

The adapter object can start functioning only when the global VM bus
channel and the corresponding per device VM bus channel are
initialized. Notifications about arrival of a virtual compute PCI
device and VM bus channels can happen in any order. Therefore,
the initial dxgadapter object state is DXGADAPTER_STATE_WAITING_VMBUS.
A list of VM bus channels and a list of waiting dxgadapter objects
are maintained. When dxgkrnl is notified about a VM bus channel
arrival, if tries to start all adapters, which are not started yet.

Properties of the adapter object are determined by sending VM
bus messages to the host to the corresponding VM bus channel.

When the per virtual compute device VM bus channel or the global
channel are destroyed, the adapter object is destroyed.

Signed-off-by: Iouri Tarassov <iourit@linux.microsoft.com>